Debug working in eclipse
Create a new run config in eclipse. C/C++ Application. Point to the exe
in devel/lib/<pckg-name>/<node-name>. Changed in the CMakeList to always
use debug build. Debug must be start witht the debug running, but no
other stuff is needed. Eclipse was started from terminal, so the
enivoronment was properly setup.